requirements.txt
certifi==2021.10.8
numpy==1.22.3
pandas==1.4.1
Pillow==9.0.1
python-dateutil==2.8.2
pytz==2022.1
pyzbar==0.1.9
qrcode==7.3.1
six==1.16.0
